Final Submit
Everything is working well except collisions between cars in the LOO
with the other cars in LOO. Adds the existing collision code. Truck AI
should be improved so it doesn’t run over the spy when approaching.
Still has the weird collision bug that causes cars to jitter (seems to
be linked to hlpr-truck).