Began conversion to racket
All parameters are properly reordered to the new vehicle struct. All
accessors have been properly switched to comply with the vehicle
struct. Fixed a bug where the truck wouldn’t move because of a syntax
errors in the code. Wrote the “vehicle->ltrb” function, but have not
implemented it into the code yet. Now everything is working as well as
it ever has.